Check the changelog of the 17/08 to know what I have changed

+++ app/src/main/java/fr/raphew/valorantclicker/onClick/OmenClick.kt
@@ -9,17 +9,17 @@ import fr.raphew.valorantclicker.activity.MainActivity
 class OmenClick(mainActivity: MainActivity) : View.OnClickListener {
 
     val mainActivity: MainActivity = mainActivity
-    var click: Int = 0
-    var coins: Int = 0
-    var clickAgents: Int = 0
+    var click: Float = 0F
+    var coins: Float = 0F
+    var clickAgents: Float = 0F
 
     override fun onClick(v: View?) {
         when (v?.id){
             R.id.img_omen -> {
                 val sharedPreferences: SharedPreferences = mainActivity.getPreferences(Context.MODE_PRIVATE)
-                click = sharedPreferences.getInt("click", 0)
-                coins = sharedPreferences.getInt("coins", 0)
-                clickAgents = sharedPreferences.getInt("clickOmen", 0)
+                click = sharedPreferences.getFloat("click", 0F)
+                coins = sharedPreferences.getFloat("coins", 0F)
+                clickAgents = sharedPreferences.getFloat("clickOmen", 0F)
                 pp()
                 save()
                 val a = "$click "
@@ -29,14 +29,14 @@ class OmenClick(mainActivity: MainActivity) : View.OnClickListener {
     }
 
     private fun save(){
-        val click: Int = this.click
-        val coins: Int = this.coins
-        val clickAgents: Int = this.clickAgents
+        val click: Float = this.click
+        val coins: Float = this.coins
+        val clickAgents: Float = this.clickAgents
         val sharedPreferences: SharedPreferences = mainActivity.getPreferences(Context.MODE_PRIVATE)
         val editor = sharedPreferences.edit()
-        editor.putInt("click", click)
-        editor.putInt("coins", coins)
-        editor.putInt("clickOmen", clickAgents)
+        editor.putFloat("click", click)
+        editor.putFloat("coins", coins)
+        editor.putFloat("clickOmen", clickAgents)
         editor.apply()
     }
